\name{renetools-package}
\alias{renetools-package}
\alias{renetools}
\docType{package}
\title{Package providing a couple of useful functions for statistics and 
plotting and data manipulation
}
\description{
The package provides some functions which are needed in everydays R-Life. They 
will shorten the amount of time needed to perform repeatedly occuring tasks. 
The package contains functions for statistics, plotting and data manipulation. 
Some of the functions are very specific for devices or sampling sides of the
TU Dresden, Institute of Hydrobiology and the Neunzehnhain Ecological Station 
as well as for the Saidenbach Reservoir. Maybe too specific functions will be
sourced out to an extra package later.
}
\details{
\tabular{ll}{
Package: \tab renetools\cr
Type: \tab Package\cr
Version: \tab 1.34.1\cr
Date: \tab 2012-06-04\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most ~~
%~~ important functions ~~
}
\author{
Maintainer: Rene Sachse <rene.sachse@uni-potsdam.de>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\keyword{helper}
%\seealso{\code{\link{limnotools}}}
%\examples{
%~~ simple examples of the most important functions ~~
%}
