\name{tzmix}
\alias{tzmix}

\title{Calculation of temperature in zmix for TS Saidenbach according
  to Lothar Paul
}

\description{
  Calculation of temperature in zmix for TS Saidenbach according to
  Lothar Paul 
}

\usage{
 tzmix(t2m)
}

\arguments{
  \item{t2m}{numeric, temperature in 2 metres depth}
}

\author{René Sachse}

\references{
  Paul L, 1985. Das thermische Regime der Talsperre Saidenbach und
  einige Beziehungen zwischen abiotischen und biotischen
  Komponenten. Dissertation, TU Dresden, Fakult\enc{ä}{ae}t Bau-,
  Wasser- und Forstwesen. 84 pp.
}

\examples{
  tzmix(17.1)
}
\keyword{mixing depth}
\keyword{zmix}
